Add endpoint for courses + refactor some util methods
